Implement auto layout exercise. Set the following constraints:
width ratio constraint to constrain the image width to 1/3 the cell width
image aspect ratio to 3:4
trailing space constraints for the labels to not clip their text
alignment constraints for the labels: top, center, and bottom respectively from the labels to the image view
>= 10 top space constraint from the description to the image bottom
bottom space constraint from the description to the superview

A few of the images are distorted when drawn with an aspect ratio of 3:4, e.g. Venus and Mars.
This is due to the AC that the aspect ratio be 3:4.